* Replaced many long, unsigned long types with uint32_t and int32_t * Moved some code to header files (thing.h and scheduler.h)

+++ otserv/trunk/account.h
@@ -32,7 +32,7 @@ public:
   Account();
   ~Account();
 
-  unsigned long accnumber;
+  uint32_t accnumber;
   int premDays;    // Premium days
 
   std::string password;
