+++ otserv/trunk/baseevents.cpp
@@ -147,7 +147,7 @@ bool Event::loadScript(const std::string& scriptFile)
 		std::cout << m_scriptInterface->getLastLuaError() << std::endl;
 		return false;
 	}
-	long id = m_scriptInterface->getEvent(getScriptEventName());
+	int32_t id = m_scriptInterface->getEvent(getScriptEventName());
 	if(id == -1){
 		std::cout << "Warning: [Event::loadScript] Event " << getScriptEventName() << " not found. " << scriptFile << std::endl;
 		return false;
@@ -184,7 +184,7 @@ bool CallBack::loadCallBack(LuaScriptInterface* _interface, std::string name)
 	
 	m_scriptInterface = _interface;
 	
-	long id = m_scriptInterface->getEvent(name);
+	int32_t id = m_scriptInterface->getEvent(name);
 	if(id == -1){
 		std::cout << "Warning: [CallBack::loadCallBack] Event " << name << " not found." << std::endl;
 		return false;
